useSquiggle and useSquigglePartial hooks
